requirements.txt
flask==3.1.3
gunicorn==23.0.0
python-dotenv==1.0.0
sendgrid==6.11.0
